+++ Turtel-App/ServerApp/DomainPrimitives/Person/Gender.cs
@@ -1,7 +1,7 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace Turtel_App.ServerApp.DomainPrimitives
+namespace Turtel_App.ServerApp.DomainPrimitives.Person
 {
     [ComplexType]
     public class Gender
