+++ Turtel-Test/TestHelper/TestSetup.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
-using RobotSwarms.HICCUP;
 using Turtel_App.ServerApp.Certification;
 
 namespace RobotSwarms.TestHelper
